🚧 ♻️ First abstraction of DD search
Add Proof.apply_theorem to make the proof state manager responsible for applying theorem and resolving dependencies of the given mapping.

Add WIP DeductiveAgent interface with BFSDeductor implementation by default (retro-compatibly), it will soon change in order to give actions to the proof directly instead of applying the theorem itself like now.
The goal will be to get closer to a gym-like agent considering the proof object as the environment.

AR should and will be compatible actions at the same level as applying theorem, being used as last resort actions by default as it was before. On the same level, asking for constructions action should be added (and taken only at last after AR exaustion).